Fix warning: implicit conversion changes signedness
visp/modules/core/src/math/matrix/vpMatrix.cpp: In member function ‘vpColVector vpMatrix::eigenValues() const’:
visp/modules/core/src/math/matrix/vpMatrix.cpp:1258:26: warning: unsigned conversion from ‘int’ to ‘unsigned int’ changes value from ‘-1’ to ‘4294967295’ [-Wsign-conversion]
 1258 |     unsigned int lwork = -1;
      |                          ^~
visp/modules/core/src/math/matrix/vpMatrix.cpp:1261:83: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
 1261 |     vpMatrix::blas_dsyev(jobz, uplo, rowNum, A.data, colNum, evalue.data, &wkopt, lwork, info);
      |                                                                                   ^~~~~
visp/modules/core/src/math/matrix/vpMatrix.cpp:1262:13: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
 1262 |     lwork = static_cast<int>(wkopt);
      |             ^~~~~~~~~~~~~~~~~~~~~~~
visp/modules/core/src/math/matrix/vpMatrix.cpp:1264:86: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
 1264 |     vpMatrix::blas_dsyev(jobz, uplo, rowNum, A.data, colNum, evalue.data, WORK.data, lwork, info);
      |                                                                                      ^~~~~
visp/modules/core/src/math/matrix/vpMatrix.cpp: In member function ‘void vpMatrix::eigenValues(vpColVector&, vpMatrix&) const’:
visp/modules/core/src/math/matrix/vpMatrix.cpp:1390:26: warning: unsigned conversion from ‘int’ to ‘unsigned int’ changes value from ‘-1’ to ‘4294967295’ [-Wsign-conversion]
 1390 |     unsigned int lwork = -1;
      |                          ^~
visp/modules/core/src/math/matrix/vpMatrix.cpp:1393:83: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
 1393 |     vpMatrix::blas_dsyev(jobz, uplo, rowNum, A.data, colNum, evalue.data, &wkopt, lwork, info);
      |                                                                                   ^~~~~
visp/modules/core/src/math/matrix/vpMatrix.cpp:1394:13: warning: conversion to ‘unsigned int’ from ‘int’ may change the sign of the result [-Wsign-conversion]
 1394 |     lwork = static_cast<int>(wkopt);
      |             ^~~~~~~~~~~~~~~~~~~~~~~
visp/modules/core/src/math/matrix/vpMatrix.cpp:1396:86: warning: conversion to ‘int’ from ‘unsigned int’ may change the sign of the result [-Wsign-conversion]
 1396 |     vpMatrix::blas_dsyev(jobz, uplo, rowNum, A.data, colNum, evalue.data, WORK.data, lwork, info);
      |                                                                                      ^~~~~

+++ modules/core/src/math/matrix/vpMatrix.cpp
@@ -1255,12 +1255,12 @@ vpColVector vpMatrix::eigenValues() const
     const char uplo = 'U';
     vpMatrix A = (*this);
     vpColVector WORK;
-    unsigned int lwork = -1;
+    int lwork = -1;
     int info = 0;
     double wkopt;
     vpMatrix::blas_dsyev(jobz, uplo, rowNum, A.data, colNum, evalue.data, &wkopt, lwork, info);
     lwork = static_cast<int>(wkopt);
-    WORK.resize(lwork);
+    WORK.resize(static_cast<unsigned int>(lwork));
     vpMatrix::blas_dsyev(jobz, uplo, rowNum, A.data, colNum, evalue.data, WORK.data, lwork, info);
   }
 #endif
@@ -1387,12 +1387,12 @@ void vpMatrix::eigenValues(vpColVector &evalue, vpMatrix &evector) const
     const char uplo = 'U';
     vpMatrix A = (*this);
     vpColVector WORK;
-    unsigned int lwork = -1;
+    int lwork = -1;
     int info = 0;
     double wkopt;
     vpMatrix::blas_dsyev(jobz, uplo, rowNum, A.data, colNum, evalue.data, &wkopt, lwork, info);
     lwork = static_cast<int>(wkopt);
-    WORK.resize(lwork);
+    WORK.resize(static_cast<unsigned int>(lwork));
     vpMatrix::blas_dsyev(jobz, uplo, rowNum, A.data, colNum, evalue.data, WORK.data, lwork, info);
     evector = A.t();
   }
